function that prints half of a string

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -8,7 +8,7 @@
  */
 void puts_half(char *str)
 {
-	int i = 0, length = 0;
+	long int i = 0, length = 0;
 
 	while (str[i++])
 		length++;
